requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.8.1=pypi_0
appdirs=1.4.3=py_1
attrs=19.3.0=py_0
black=19.10b0=py37_0
blas=1.0=mkl
ca-certificates=2019.10.16=0
cachetools=3.1.1=pypi_0
certifi=2019.9.11=py37_0
cffi=1.13.1=py37h2e261b9_0
chardet=3.0.4=pypi_0
click=7.0=py_0
cudatoolkit=10.1.243=h6bb024c_0
entrypoints=0.3=py37_0
expect=5.45=1
flake8=3.7.9=py37_0
freetype=2.9.1=h8a8886c_1
google-auth=1.7.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
grpcio=1.25.0=pypi_0
idna=2.8=pypi_0
intel-openmp=2019.4=243
jpeg=9b=h024ee3a_2
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
markdown=3.1.1=pypi_0
mccabe=0.6.1=py37_1
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mypy_extensions=0.4.3=py37_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py37hfd86e86_0
numpy=1.17.3=py37hd14ec0e_0
numpy-base=1.17.3=py37hde5b4d6_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
openssl=1.1.1d=h7b6447c_3
pathspec=0.6.0=py_0
pillow=6.2.1=py37h34e0f95_0
pip=19.3.1=py37_0
protobuf=3.10.0=pypi_0
pyasn1=0.4.7=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycodestyle=2.5.0=py37_0
pycparser=2.19=py37_0
pyflakes=2.1.1=py37_0
python=3.7.5=h0371630_0
pytorch=1.3.1=py3.7_cuda10.1.243_cudnn7.6.3_0
readline=7.0=h7b6447c_5
regex=2019.11.1=py37h516909a_0
requests=2.22.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
setuptools=41.6.0=py37_0
six=1.12.0=py37_0
sqlite=3.30.1=h7b6447c_0
tensorboard=2.0.1=pypi_0
tk=8.6.8=hbc83047_0
toml=0.10.0=py_0
torchvision=0.4.2=py37_cu101
typed-ast=1.4.0=py37h516909a_0
typing_extensions=3.7.4.1=py37_0
urllib3=1.25.7=pypi_0
werkzeug=0.16.0=pypi_0
wheel=0.33.6=py37_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
